Move storages to seaprate files
